updated the controller with new functions

+++ Backend/controllers/message.controller.js
@@ -16,6 +16,14 @@ const sendMessage = asyncHandler(async (req, res) => {
         conversation = await Conversation.findById(conversationId);
     }
 
+    if (conversation && conversation.deletedFor.includes(senderId.toString())) {
+        conversation = await Conversation.findByIdAndUpdate(conversationId, {
+            $pull: {
+                deletedFor: senderId
+            },
+        }, { new: true });
+    }
+
     if (!conversation) {
         if (senderId.toString() === recieverId.toString()) {
             conversation = await Conversation.findOne({
@@ -24,9 +32,13 @@ const sendMessage = asyncHandler(async (req, res) => {
             })
         } else {
             conversation = await Conversation.findOne({
-                participants: { $all: [senderId, recieverId] }
+                participants: { $all: [senderId, recieverId] },
+                $expr: { $eq: [{ $size: "$participants" }, 2] }
             })
         }
+    }
+
+    if (!conversation) {
         conversation = await Conversation.create({
             participants: [senderId, recieverId]
         })
@@ -47,8 +59,8 @@ const sendMessage = asyncHandler(async (req, res) => {
     await conversation.messages.push(newMessage._id);
 
     await Promise.all([
-        conversation.save(),
-        newMessage.save()
+        conversation.save({ timeStamps: true }),
+        newMessage.save({ timeStamps: true })
     ])
 
     const sentMessage = newMessage;
@@ -61,7 +73,11 @@ const sendMessage = asyncHandler(async (req, res) => {
 const getAllMessages = asyncHandler(async (req, res) => {
     const { id: conversationId } = req.params;
 
-    const conversation = await Conversation.findById(conversationId).populate("messages")
+    const conversation = await Conversation.findById(conversationId,{
+        createdAt: {
+            $gt: clearedAt || Date.now(0)
+        }
+    }).populate("messages")
 
     return res.status(200).json(
         new ApiResponses(200, conversation, "All messages fetched successfully")
@@ -94,6 +110,11 @@ const deleteMessage = asyncHandler(async (req, res) => {
                     }
                 })
             } else {
+                await Conversation.findByIdAndUpdate(message.conversationId, {
+                    $pull: {
+                        messages: message?._id
+                    }
+                })
                 await Message.findByIdAndDelete(file.messagedId);
             }
         }
@@ -118,7 +139,23 @@ const deleteConversation = asyncHandler(async (req, res) => {
         throw new ApiErrors(401, "You are not authorized to delete this conversation");
     }
 
-    await Conversation.findByIdAndDelete(conversationId);
+    if (conversation.deletedFor.includes(req.user?._id)) {
+        throw new ApiErrors(409, "Conversation already deleted for you");
+    }
+
+    if (conversation.deletedFor.length === conversation.participants.length) {
+        await Conversation.findByIdAndDelete(conversationId);
+    } else {
+        await Conversation.findByIdAndUpdate(conversationId, {
+            $addToSet: {
+                deletedFor: req.user?._id
+            },
+            $set: {
+                [`clearedAt.${req.user?._id.toString()}`]: new Date()
+            }
+        }, { new: true }
+        );
+    }
 
     return res.status(200).json(
         new ApiResponses(200, {}, "Conversation deleted successfully")
@@ -126,4 +163,37 @@ const deleteConversation = asyncHandler(async (req, res) => {
 
 })
 
-export { sendMessage, getAllMessages, deleteMessage, deleteConversation };
+
+const editMessageContent = asyncHandler(async (req, res) => {
+
+    const { messageId } = req.params;
+    const { message } = req.body;
+
+    if (!messageId) {
+        throw new ApiErrors(404, "messageId is required!");
+    }
+
+    if (!message) {
+        throw new ApiErrors(404, "message cannot be empty!");
+    }
+
+    const isexists = await Message.findById(messageId);
+
+    if (!isexists) {
+        throw new ApiErrors(401, "message does not exsits");
+    }
+
+    const updatedMessage = await Message.findByIdAndUpdate(messageId, {
+        $set: {
+            message: message
+        }
+    })
+
+    return res.status(200)
+        .json(
+            200, updatedMessage, "message was updated successfully"
+        )
+
+})
+
+export { sendMessage, getAllMessages, deleteMessage, deleteConversation, editMessageContent };
